pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "model_arithmetic"
version = "1.0.0"
authors = [
  { name="Anonymous", email="anonymous@anonymous.com" },
]
description = "Prompt Arithmetic Package"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "accelerate >= 0.21.0",
    "loguru >= 0.7.0",
    "matplotlib >= 3.7.2",
    "numpy >= 1.24.3",
    "pandas >= 2.0.3",
    "trl >= 0.5.0",
    "transformers >= 4.31.0",
    "torch >= 2.0.1",
    "scikit-learn >= 1.3.0",
    "sentencepiece >= 0.1.99",
    "peft >= 0.4.0", 
    "fuzzywuzzy >= 0.18.0",
    "google-api-python-client >= 2.97.0",
    "python-Levenshtein >= 0.21.1",
    "python_dotenv >= 1.0.0",
    "charset_normalizer>=3.2.0",
    "aiohttp >= 3.8.5"
]
